+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -4,7 +4,7 @@ Release History
 ===============
 0.1.8 (unreleased)
 ++++++++++++++++++++
-webapp: expose "production-test" command to configure static routing
+webapp: expose "traffic-routing" commands to configure static routing
 
 0.1.7 (2017-05-30)
 ++++++++++++++++++++